+++ ADC_TRABALHO/src/Classes/__pycache__/funcionario.cpython-312.pyc
