+++ ADC_TRABALHO/src/Classes/__pycache__/funcionario.cpython-313.pyc
