+++ ADC_TRABALHO/src/Classes/leitor.py
